requirements.txt
flask
flask-restx
flask-sqlalchemy
werkzeug
python-dotenv
flask-jwt-extended
PyJWT